% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beast}
\alias{beast}
\title{beast}
\usage{
beast(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run beast
}
\examples{
library(outsider)
beast <- module_import('beast', repo = 'dombennett/om..beast')
beast('-help')
}
